Refactor register page
