+++ web/src/pages/register/components/RegisterFormFooter.tsx
@@ -1,0 +1,19 @@
+import { Link } from 'react-router-dom';
+import { Button } from '@/components/ui/button';
+import { CardFooter } from '@/components/ui/card';
+
+export default function RegisterFormFooter({ isLoading }: { isLoading: boolean }) {
+  return (
+    <CardFooter className="flex flex-col gap-4">
+      <Button type="submit" className="w-full" disabled={isLoading}>
+        {isLoading ? 'Creating account...' : 'Create account'}
+      </Button>
+      <p className="text-sm text-center text-muted-foreground">
+        Already have an account?{' '}
+        <Link to="/login" className="text-primary hover:underline font-medium">
+          Sign in
+        </Link>
+      </p>
+    </CardFooter>
+  );
+}
